IChatService, ChatService and Chat, Filters, modifications added to Host, InternalInfoUser and ExternalInfoUser. tests added. DB´Tables added for filters  added

+++ GymApp/Data Access Layer/Shared/DBCommand.cs
@@ -44,7 +44,7 @@ namespace Data_Access_Layer.Shared
             }
         }
 
-        public void AddQueryParamters(string key, float value)
+        public void AddQueryParamters(string key, double value)
         {
             IDataParameter param = DBComponentsFactory.ComponentProvider.CreateParameter();
             param.ParameterName = key;
